etl/requirements.txt
boto3==1.37.31
pyspark==3.5.5
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sqlalchemy==2.0.40
psycopg2-binary==2.9.10